fix test + move check inside backend
