+++ crates/typos-cli/Cargo.toml
@@ -85,7 +85,7 @@ regex = "1.7.2"
 tracing-subscriber = { version = "0.3.16", default-features = false, features = ["env-filter", "fmt", "smallvec", "std"] }
 tower-lsp = "0.19.0"
 tracing = "0.1.37"
-tokio = { version = "1.27.0", features = ["macros", "rt-multi-thread", "io-std"] }
+tokio = { version = "1.27.0", features = ["macros", "rt-multi-thread", "io-std", "io-util"] }
 
 [dev-dependencies]
 assert_fs = "1.0"
